asignatura listo

+++ gestion-app/app/asignaturas/page.tsx
@@ -30,10 +30,13 @@ export default function Asignaturas() {
     credits: 0,
     semester_id: 0,
   });
+  
   const [isEditing, setIsEditing] = useState<boolean>(false);
   const [showGeneralInfo, setShowGeneralInfo] = useState<boolean>(false);
   const [selectedSemester, setSelectedSemester] = useState<number | null>(null); // Nuevo estado para el filtro de semestre
 
+  
+  
   // Obtener las asignaturas y los semestres de la base de datos
   useEffect(() => {
     async function fetchData() {
@@ -83,22 +86,42 @@ export default function Asignaturas() {
     setNewAsignatura({ course_id: 0, course_name: '', code: '', credits: 0, semester_id: 0 }); // Resetear el formulario
   };
 
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
     const { name, value } = e.target;
-    if (selectedAsignatura) {
-      setSelectedAsignatura((prev) => ({
-        ...prev!,
-        [name]: value,
+    
+    if (name === 'credits') {
+      setNewAsignatura((prev) => ({
+        ...prev,
+        [name]: Number(value), // Convert to number for credits
+      }));
+    } else if (name === 'semester_id') {
+      setNewAsignatura((prev) => ({
+        ...prev,
+        [name]: value ? Number(value) : 0, // Ensure semester_id is a number
       }));
     } else {
-      setNewAsignatura((prev) => ({ ...prev, [name]: value }));
+      setNewAsignatura((prev) => ({
+        ...prev,
+        [name]: value,
+      }));
     }
   };
+  
+  
+  
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    if (newAsignatura.course_name && newAsignatura.code && newAsignatura.credits) {
+    
+    // Verificar si los campos tienen valores válidos
+    if (
+      newAsignatura.course_name.trim() !== "" &&
+      newAsignatura.code.trim() !== "" &&
+      newAsignatura.credits != 0 && 
+      newAsignatura.semester_id !==0
+    ) {
       try {
+        // Hacer una solicitud POST para agregar la nueva asignatura
         const response = await fetch('/api/asignaturas', {
           method: 'POST',
           headers: {
@@ -106,12 +129,15 @@ export default function Asignaturas() {
           },
           body: JSON.stringify(newAsignatura),
         });
+        
         if (!response.ok) {
           throw new Error('Error al agregar la asignatura');
         }
+  
         const addedAsignatura = await response.json();
         setAsignaturas((prev) => [...prev, addedAsignatura]);
         handleModalClose();
+        
       } catch (error) {
         console.error(error);
         alert('Hubo un problema al agregar la asignatura.');
@@ -120,56 +146,43 @@ export default function Asignaturas() {
       alert("Por favor, completa todos los campos.");
     }
   };
+  
+  
+  
+// En tu componente de asignaturas
+const handleDelete = async (id: number) => {
+  // Confirmación antes de eliminar
+  const confirmation = window.confirm('¿Estás seguro de que quieres eliminar esta asignatura?');
+  
+  if (confirmation) {
+    try {
+      // Realizar la solicitud DELETE al backend
+      const response = await fetch(`/api/asignaturas?course_id=${id}`, {
+        method: 'DELETE',
+      });
 
-  const handleEliminarAsignatura = async (id: number) => {
-    await fetch(`/api/asignaturas/${id}`, {
-      method: 'DELETE',
-    });
-    setAsignaturas(asignaturas.filter(asignatura => asignatura.course_id !== id));
-    handleClose(); // Cierra el detalle si la asignatura eliminada era la seleccionada
-  };
-
-  const handleModificar = () => {
-    setIsEditing(true); // Habilitar el modo de edición
-  };
-
-  const handleActualizar = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (selectedAsignatura) {
-      try {
-        const response = await fetch(`/api/asignaturas`, {
-          method: 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(selectedAsignatura),
-        });
-        if (!response.ok) {
-          throw new Error('Error al actualizar la asignatura');
-        }
-        const updatedAsignatura = await response.json();
-        setAsignaturas((prev) =>
-          prev.map((asignatura) =>
-            asignatura.course_id === updatedAsignatura.course_id ? updatedAsignatura : asignatura
-          )
-        );
-        setIsEditing(false);
-      } catch (error) {
-        console.error(error);
-        alert('Hubo un problema al actualizar la asignatura.');
+      if (!response.ok) {
+        throw new Error('Error al eliminar la asignatura');
       }
+
+      setAsignaturas((prevAsignaturas) =>
+        prevAsignaturas.filter((asignatura) => asignatura.course_id !== id)
+      );
+      
+      alert('Asignatura eliminada con éxito');
+      
+    } catch (error) {
+      console.error(error);
+      alert('Hubo un problema al eliminar la asignatura.');
     }
-  };
+  }
+}
 
-  const handleVerInfoGeneral = () => {
-    setShowGeneralInfo(true);
-  };
 
-  const handleCloseGeneralInfo = () => {
-    setShowGeneralInfo(false);
-  };
+  
 
   return (
+    
     <div className="layout asignaturas-page">
       {/* Filtro por semestre */}
       <label htmlFor="semester-filter">Filtrar por Semestre:</label>
@@ -181,71 +194,113 @@ export default function Asignaturas() {
           </option>
         ))}
       </select>
-
-      <button onClick={handleVerInfoGeneral}>Ver Info General</button>
-
-      {showGeneralInfo && (
-        <div className="modal-overlay">
-          <div className="modal">
-            <div className="modal-header">
-              <h2>Informe General de Asignaturas</h2>
-              <button className="close-button" onClick={handleCloseGeneralInfo}>X</button>
-            </div>
-            <div className="modal-content">
-              <div className="grid">
-                {filteredAsignaturas.map((asignatura) => (
-                  <div className="grid-item" key={asignatura.course_id}>
-                    <h3>{asignatura.course_name}</h3>
-                    <p>Código: {asignatura.code}</p>
-                    <p>Créditos: {asignatura.credits}</p>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
-
+        
       <div className="sidebar">
-        <h2 className="sidebar-title">Asignaturas Disponibles</h2>
-        <button className="add-button" onClick={handleAgregarAsignatura}>
-          Agregar Asignatura
+  <h2 className="sidebar-title">Asignaturas Disponibles</h2>
+  <button className="add-button" onClick={handleAgregarAsignatura}>
+    Agregar Asignatura
+  </button>
+  
+  <ul className="asignaturas-list">
+    {filteredAsignaturas.map((asignatura) => (
+      <li key={asignatura.course_id} className="asignatura-item">
+        <span className="asignatura-info">
+          {asignatura.course_name} - {asignatura.code} - Créditos: {asignatura.credits}
+        </span>
+        <button
+          className="info-button"
+          onClick={() => handleVerInformacion(asignatura)}
+        >
+          Ver Información
         </button>
-        <ul className="asignaturas-list">
-          {filteredAsignaturas.map((asignatura) => (
-            <li key={asignatura.course_id} className="asignatura-item">
-              <span className="asignatura-info">
-                {asignatura.course_name} - {asignatura.code} - Créditos: {asignatura.credits}
-              </span>
-              <button className="info-button" onClick={() => handleVerInformacion(asignatura)}>
-                Ver Información
-              </button>
-            </li>
+      </li>
+    ))}
+  </ul>
+</div>
+{isModalOpen && (
+  <div className="modal">
+    <div className="modal-content">
+      <h2>Agregar Nueva Asignatura</h2>
+      <form onSubmit={handleSubmit}>
+        <label htmlFor="course_name">Nombre de la Asignatura:</label>
+        <input
+          type="text"
+          id="course_name"
+          name="course_name"
+          value={newAsignatura.course_name}
+          onChange={handleInputChange}
+        />
+        <label htmlFor="code">Código:</label>
+        <input
+            type="text"
+            id="code"
+            name="code"
+            value={newAsignatura.code}
+            onChange={handleInputChange}
+/>
+
+        <label htmlFor="credits">Créditos:</label>
+        <input
+          type="number"
+          id="credits"
+          name="credits"
+          value={newAsignatura.credits}
+          onChange={handleInputChange}
+          min="1" 
+          required
+        />
+        
+        <label htmlFor="semester_id">Semestre:</label>
+        <select
+          id="semester_id"
+          name="semester_id"
+          value={newAsignatura.semester_id}
+          onChange={handleInputChange}
+        >
+          <option value="">Seleccionar Semestre</option>
+          {semestres.map(semester => (
+            <option key={semester.semester_id} value={semester.semester_id}>
+              {semester.semester_num}
+            </option>
           ))}
-        </ul>
+        </select>
+
+        <div className="modal-actions">
+          <button type="submit">Guardar</button>
+          <button type="button" onClick={handleModalClose}>
+            Cancelar
+          </button>
+        </div>
+      </form>
+    </div>
+  </div>
+)}
+
+
+{/* Información Detallada de la Asignatura */}
+<div className="details">
+  {selectedAsignatura && (
+    <div className="details-card">
+      <h2 className="details-title">Información de {selectedAsignatura.course_name}</h2>
+      <p>Código: {selectedAsignatura.code}</p>
+      <p>Créditos: {selectedAsignatura.credits}</p>
+      <p>Semestre: {semestres.find(semester => semester.semester_id === selectedAsignatura.semester_id)?.semester_name}</p>
+      <div className="details-menu">
+        <button
+          className="details-button"
+          onClick={() => handleDelete(selectedAsignatura.course_id)}
+        >
+          Eliminar 
+        </button>
+        
+        <button className="details-button" onClick={handleClose}>
+          Cerrar
+        </button>
       </div>
-      {/* Información Detallada de la Asignatura */}
-      <div className="details">
-        {selectedAsignatura && (
-          <div className="details-card">
-            <h2 className="details-title">Información de {selectedAsignatura.course_name}</h2>
-            <p>Código: {selectedAsignatura.code}</p>
-            <p>Créditos: {selectedAsignatura.credits}</p>
-            <p>Semestre: {semestres.find(semester => semester.semester_id === selectedAsignatura.semester_id)?.semester_name}</p>
-            {isEditing ? (
-              <div>
-                <button className="edit-button" onClick={handleActualizar}>Actualizar</button>
-              </div>
-            ) : (
-              <button className="edit-button" onClick={handleModificar}>Modificar</button>
-            )}
-            <button className="delete-button" onClick={() => handleEliminarAsignatura(selectedAsignatura.course_id)}>
-              Eliminar
-            </button>
-            <button className="close-button" onClick={handleClose}>Cerrar</button>
-          </div>
-        )}
-      </div>
+    </div>
+  )}
+</div>
+
     </div>
   );
 }
